refactored summaries controller and test to set confirmation code and transmit application with consent
