+++ app/app/controllers/cbv/successes_controller.rb
@@ -1,38 +1,6 @@
 class Cbv::SuccessesController < Cbv::BaseController
   skip_before_action :ensure_cbv_flow_not_yet_complete
-  before_action :transmit_to_caseworker
 
   def show
-  end
-
-  private
-
-  def transmit_to_caseworker
-    if @cbv_flow.confirmation_code.blank?
-      confirmation_code = generate_confirmation_code(@cbv_flow.site_id)
-      @cbv_flow.update(confirmation_code: confirmation_code)
-    end
-
-    if current_site.transmission_method.present? && current_site.transmission_method == "shared_email"
-      CaseworkerMailer.with(
-        email_address: current_site.transmission_method_configuration.dig("email"),
-        cbv_flow: @cbv_flow,
-        payments: @payments
-      ).summary_email.deliver_now
-      @cbv_flow.touch(:transmitted_at)
-    end
-
-    NewRelicEventTracker.track("IncomeSummarySharedWithCaseworker", {
-      timestamp: Time.now.to_i,
-      cbv_flow_id: @cbv_flow.id
-    })
-  end
-
-  def generate_confirmation_code(prefix = nil)
-    [
-      prefix,
-      (Time.now.to_i % 36 ** 3).to_s(36).tr("OISB", "0158").rjust(3, "0"),
-      @cbv_flow.id.to_s.rjust(4, "0")
-    ].compact.join.upcase
   end
 end
