+++ app/spec/controllers/cbv/summaries_controller_spec.rb
@@ -2,29 +2,18 @@ require "rails_helper"
 
 RSpec.describe Cbv::SummariesController do
   include PinwheelApiHelper
-  let(:cbv_flow_props) do  {
-    first_name: "John",
-    middle_name: "Doe",
-    last_name: "Smith",
-    case_number: "ABC1234",
-    email_address: "tom@example.com",
-    agency_id_number: "A12345",
-    site_id: "sandbox",
-    snap_application_date: Date.parse('2024-06-18'),
-    created_at: Time.new(2024, 8, 1, 12, 0, 0, "-04:00"),
-    id: 1
-  }
-  end
-  let(:cbv_flow_invitation) { CbvFlowInvitation.create(cbv_flow_props) }
-  let(:cbv_flow) { CbvFlow.create(case_number: "ABC1234", pinwheel_token_id: "abc-def-ghi", site_id: "sandbox", cbv_flow_invitation_id: 1) }
+
+  let(:cbv_flow) { create(:cbv_flow, case_number: "ABC1234", pinwheel_token_id: "abc-def-ghi") }
+  let(:cbv_flow_invitation) { cbv_flow.cbv_flow_invitation }
 
   before do
     session[:cbv_flow_invitation] = cbv_flow_invitation
-    session[:cbv_flow] = cbv_flow
   end
 
   describe "#show" do
     before do
+      cbv_flow_invitation.update(snap_application_date: Date.parse('2024-06-18'))
+      cbv_flow_invitation.update(created_at: Date.parse('2024-03-20'))
       session[:cbv_flow_id] = cbv_flow.id
       stub_request_end_user_accounts_response
       stub_request_end_user_paystubs_response
@@ -35,7 +24,7 @@ RSpec.describe Cbv::SummariesController do
 
       it "renders properly" do
         get :show
-        expect(assigns[:already_consented]).to eq(false)
+        expect(controller.send(:has_consent)).to be_falsey
         # 90 days before snap_application_date
         start_date = "March 20, 2024"
         # Should be the formatted version of snap_application_date
@@ -66,15 +55,10 @@ RSpec.describe Cbv::SummariesController do
     end
 
     context "for a completed CbvFlow" do
-      let(:cbv_flow) do
-        CbvFlow.create(
-          case_number: "ABC1234",
-          pinwheel_token_id: "abc-def-ghi",
-          site_id: "sandbox",
-          confirmation_code: "ABC1234",
-          cbv_flow_invitation_id: 1
-        )
+      before do
+        cbv_flow.update(confirmation_code: "ABC123")
       end
+
       it "allows the user to download the PDF summary" do
         get :show, format: :pdf
         expect(response).to be_successful
@@ -84,6 +68,77 @@ RSpec.describe Cbv::SummariesController do
       it "redirects the user to the success page if the user goes back to the page" do
         get :show
         expect(response).to redirect_to(cbv_flow_success_path)
+      end
+    end
+  end
+
+  describe "#update" do
+    let(:nyc_user) { User.create(email: "test@test.com", site_id: 'nyc') }
+
+    before do
+      session[:cbv_flow_id] = cbv_flow.id
+      sign_in nyc_user
+      stub_request_end_user_accounts_response
+      stub_request_end_user_paystubs_response
+    end
+
+    context "without consent" do
+      it "redirects back with an alert" do
+        patch :update
+        expect(response).to redirect_to(cbv_flow_summary_path)
+        expect(flash[:alert]).to be_present
+        expect(flash[:alert]).to eq("You must check the legal agreement checkbox to proceed.")
+      end
+    end
+
+    context "with consent" do
+      it "generates a new confirmation code" do
+        expect(cbv_flow.confirmation_code).to be_nil
+        patch :update, params: { cbv_flow: { consent_to_authorized_use: "1" }, token: cbv_flow_invitation.auth_token }
+        cbv_flow.reload
+        expect(cbv_flow.confirmation_code).to start_with("SANDBOX")
+      end
+    end
+
+    context "when confirmation_code already exists" do
+      let(:existing_confirmation_code) { "SANDBOX000" }
+
+      before do
+        cbv_flow.update(confirmation_code: existing_confirmation_code)
+      end
+
+      it "does not override the existing confirmation code" do
+        expect(cbv_flow.reload.confirmation_code).to eq(existing_confirmation_code)
+        expect { patch :update }.not_to change { cbv_flow.reload.confirmation_code }
+      end
+    end
+
+    context "when sending an email to the caseworker" do
+      before do
+        cbv_flow.update(consented_to_authorized_use_at: Time.now)
+        stub_request_end_user_accounts_response
+        stub_request_end_user_paystubs_response
+      end
+
+      it "sends the email" do
+        expect do
+          patch :update
+        end.to change { ActionMailer::Base.deliveries.count }.by(1)
+
+        email = ActionMailer::Base.deliveries.last
+        expect(email.subject).to eq("Applicant Income Verification: ABC1234")
+      end
+
+      it "stores the current time as transmitted_at" do
+        expect { patch :update }
+          .to change { cbv_flow.reload.transmitted_at }
+                .from(nil)
+                .to(within(5.second).of(Time.now))
+      end
+
+      it "redirects to success screen" do
+        patch :update
+        expect(response).to redirect_to({ controller: :successes, action: :show })
       end
     end
   end
